refactor: extract shared data hook + presentational helpers from PanelViews.tsx
Foundation PR for issue #144 (PanelViews.tsx split into per-panel files).
No panel bodies moved yet — pure plumbing swap so every subsequent batch
PR can extract panels independently against a stable shared-data layer.

- New panels/usePanelData.ts: the 6 apiFetch calls + useEffects, fallback
  datasets, students/schools/usersList/reportsList/worksheetsList/
  teachersList derivations, getDistrictStats/getBlockStats — moved
  verbatim from PanelViews.tsx.
- New panels/PanelShared.tsx: PageHeader + EmptyStudents, moved verbatim
  (now exported for reuse by future per-panel files).
- New panels/pdfReportGenerator.ts: handleDownloadPDF, moved verbatim
  (pure function, no closure over component state).
- One intentional behavior-preserving change: the hook exposes a named
  updateStudentLocally(studentId, patch) mutator instead of leaking
  setApiStudents — the only shared-data mutation in the file (student_profile's
  optimistic update after PATCH /api/students/:id/profile).

Verified: tsc --noEmit clean, vite build succeeds, and a live click-through
against a scratch-seeded backend as Superadmin/State Admin/Teacher —
including an actual profile edit through updateStudentLocally, confirmed
via a direct MongoDB read that the PATCH persisted correctly.

While verifying, found and filed #245: a pre-existing, unrelated hooks-order
crash in the content panel (reproduces identically on unmodified main) —
left unfixed here per the no-behavior-change scope of this PR; will be
naturally resolved when content is extracted to its own file later in
this same split sequence.
